hotpotqa_runs/requirements.txt
gym==0.26.2
joblib==1.2.0
langchain==0.0.162
numpy==1.24.1
openai==0.27.4
python-dotenv==1.0.0
tenacity==8.2.2
tiktoken==0.4.0
transformers==4.28.1
pandas==1.5.3
scikit-learn
wikipedia